# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "API documentation for Online Retail API",
      "version": "1.0.0",
      "title": "Online Retail API"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
      "/products": {
        "get": {
          "summary": "Get all products",
          "responses": {
            "200": {
              "description": "Successful response"
            }
          }
        },
        "post": {
          "summary": "Add a new product",
          "parameters": [
            {
              "in": "body",
              "name": "product",
              "description": "Product to add",
              "schema": {
                "$ref": "#/definitions/Product"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Product added"
            }
          }
        }
      },
      "/products/{id}": {
        "get": {
          "summary": "Get a specific product by ID",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Product found"
            },
            "404": {
              "description": "Product not found"
            }
          }
        },
        "put": {
          "summary": "Update an existing product",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "product",
              "description": "Product to update",
              "schema": {
                "$ref": "#/definitions/Product"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Product updated"
            },
            "404": {
              "description": "Product not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a product",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Product deleted"
            },
            "404": {
              "description": "Product not found"
            }
          }
        }
      }
    },
    "definitions": {
      "Product": {
        "type": "object",
        "properties": {
          "invoiceno": {
            "type": "string"
          },
          "stockcode": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "invoicedate": {
            "type": "string",
            "format": "date-time"
          },
          "unitprice": {
            "type": "number",
            "format": "float"
          },
          "customerid": {
            "type": "integer"
          },
          "country": {
            "type": "string"
          }
        }
      }
    }
  }
  